Код отформатирован и добавлена валидация токена сервера vk
