+++ src/main/java/ru/mityushin/responder/ResponderApplication.java
@@ -3,10 +3,6 @@ package ru.mityushin.responder;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-/**
- * @author Dmitry Mityushin
- * @version 1.0
- */
 @SpringBootApplication
 public class ResponderApplication {
