--- src/main/java/ru/mityushin/responder/commandsmodule/Command.java
@@ -1,80 +1,0 @@
-package ru.mityushin.responder.commandsmodule;
-
-import com.vk.api.sdk.objects.messages.Message;
-import ru.mityushin.responder.service.VkMessageSenderService;
-
-/**
- * Abstract class for all executable classes-commands
- * Field {@link #name} identification command,he is called by this name
- *
- * @author Arthur Kupriyanov
- * @version 1.1
- */
-public abstract class Command {
-    protected final VkMessageSenderService messageSenderService;
-
-    public final String name;
-
-    public Command(VkMessageSenderService messageSenderService, String name){
-        this.messageSenderService = messageSenderService;
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getParam() {
-        return param;
-    }
-
-    public void setParam(String param) {
-        this.param = param;
-    }
-
-    private String param;
-
-    /**
-     * Метод, который будет вызываться для исполнения команды
-     * @param message сообщение пользователя
-     */
-    public abstract void exec(Message message);
-
-    /**
-     * Возвращает строку в формате:<br>
-     * name: имяКоманды<br>
-     *
-     * @return форматированное имя и мод команды
-     */
-    @Override
-    public String toString() {
-        return String.format("name: %s",this.name);
-    }
-
-    /**
-     * Берет хэш-код значащего поля {@link #name}
-     *
-     * @return хэш-код команды
-     */
-    @Override
-    public int hashCode() {
-        return this.name.hashCode();
-    }
-
-
-    /**
-     * Объекты эквивалентны только, если поля <code>{@link #name}</code> равны
-     * имеют одинаковое значение и объект является классом-наследником {@link Command}
-     * @param obj сравниваемый объект
-     * @return {@code true} если объекты эквивалентны; {@code false} если объекты различаются
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (obj instanceof Command){
-            if (name.equals(((Command) obj).name)){
-                return true;
-            }
-        }
-        return false;
-    }
-}
