+++ src/main/java/ru/mityushin/responder/commandsmodule/commands/Command.java
@@ -1,0 +1,51 @@
+package ru.mityushin.responder.commandsmodule.commands;
+
+import com.vk.api.sdk.objects.messages.Message;
+import ru.mityushin.responder.service.VkMessageSenderService;
+
+public abstract class Command {
+    protected final VkMessageSenderService messageSenderService;
+
+    public final String name;
+
+    public Command(VkMessageSenderService messageSenderService, String name){
+        this.messageSenderService = messageSenderService;
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getParam() {
+        return param;
+    }
+
+    public void setParam(String param) {
+        this.param = param;
+    }
+
+    private String param;
+
+    public abstract void exec(Message message);
+
+    @Override
+    public String toString() {
+        return String.format("name: %s",this.name);
+    }
+
+    @Override
+    public int hashCode() {
+        return this.name.hashCode();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof Command){
+            if (name.equals(((Command) obj).name)){
+                return true;
+            }
+        }
+        return false;
+    }
+}
