+++ src/main/java/ru/mityushin/responder/commandsmodule/commands/Help.java
@@ -1,13 +1,10 @@
-package ru.mityushin.responder.commandsmodule;
+package ru.mityushin.responder.commandsmodule.commands;
 
 import com.vk.api.sdk.exceptions.ApiException;
 import com.vk.api.sdk.exceptions.ClientException;
 import com.vk.api.sdk.objects.messages.Message;
 import ru.mityushin.responder.service.VkMessageSenderService;
 
-/**
- * @author Arthur Kupriyanov
- */
 public class Help extends Command {
 
     public Help(VkMessageSenderService messageSenderService, String name) {
