+++ src/main/java/ru/mityushin/responder/commandsmodule/commands/Play.java
@@ -1,26 +1,22 @@
-package ru.mityushin.responder.commandsmodule;
-
+package ru.mityushin.responder.commandsmodule.commands;
 
 import com.vk.api.sdk.exceptions.ApiException;
 import com.vk.api.sdk.exceptions.ClientException;
 import com.vk.api.sdk.objects.messages.Message;
+import ru.mityushin.responder.checkers.NewCheckersBoard;
 import ru.mityushin.responder.service.VkMessageSenderService;
 
-/**
- * @author Arthur Kupriyanov
- */
-public class Unknown extends Command {
+public class Play extends Command {
 
-    public Unknown(VkMessageSenderService messageSenderService, String name) {
+    public Play(VkMessageSenderService messageSenderService, String name) {
         super(messageSenderService, name);
     }
 
     @Override
     public void exec(Message message) {
-        //new VKManager().sendMessage("Неизвестная команда", message.getUserId());
-        message.setText("Unknown command, bro!");
+        NewCheckersBoard checkersBoard = NewCheckersBoard.getCheckersBoard();
         try {
-            messageSenderService.send(message);
+            messageSenderService.sendPhoto(message, checkersBoard.getBoard());
         } catch (ClientException e) {
             e.printStackTrace();
         } catch (ApiException e) {
