+++ src/main/java/ru/mityushin/responder/commandsmodule/commands/Reset.java
@@ -1,11 +1,10 @@
-package ru.mityushin.responder.commandsmodule;
+package ru.mityushin.responder.commandsmodule.commands;
 
 import com.vk.api.sdk.exceptions.ApiException;
 import com.vk.api.sdk.exceptions.ClientException;
 import com.vk.api.sdk.objects.messages.Message;
 import ru.mityushin.responder.checkers.NewCheckersBoard;
 import ru.mityushin.responder.service.VkMessageSenderService;
-
 import java.io.IOException;
 
 public class Reset extends Command {
