+++ src/main/java/ru/mityushin/responder/controller/CallbackController.java
@@ -10,20 +10,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import ru.mityushin.responder.config.VKGroupActor;
 import ru.mityushin.responder.service.CallbackApiHandler;
 
-/**
- * Controller for VK API requests
- *
- * @author Dmitry Mityushin
- * @since 1.0
- */
 @Controller
 @RequestMapping(value = "/callbacks")
 @RequiredArgsConstructor
 public class CallbackController {
     private static final Logger LOG = LoggerFactory.getLogger(CallbackController.class);
     private final CallbackApiHandler callbackApiHandler;
+    private final VKGroupActor vkGroupActor;
 
     @PostMapping
     @ResponseBody
@@ -33,7 +29,7 @@ public class CallbackController {
         JsonObject json = gson.fromJson(callbackDto, JsonObject.class);
         String type = json.get("type").getAsString();
         if (type.equalsIgnoreCase("confirmation")) {
-            return "d70aff5a";
+            return vkGroupActor.getConfirmation();
         }else {
             callbackApiHandler.parse(callbackDto);
             return "ok";
