+++ src/main/java/ru/mityushin/responder/service/CallbackApiHandler.java
@@ -3,49 +3,36 @@ package ru.mityushin.responder.service;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.vk.api.sdk.callback.CallbackApi;
-import com.vk.api.sdk.exceptions.ApiException;
-import com.vk.api.sdk.exceptions.ClientException;
 import com.vk.api.sdk.objects.callback.ConfirmationMessage;
 import com.vk.api.sdk.objects.messages.Message;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.mityushin.responder.commandsmodule.Commander;
-import ru.mityushin.responder.controller.CallbackController;
+import ru.mityushin.responder.config.VKGroupActor;
 
 @Service
 public class CallbackApiHandler extends CallbackApi {
-    private static final Logger LOG = LoggerFactory.getLogger(CallbackController.class);
-    private final VkMessageSenderService vkMessageSenderService;
+    private final VKGroupActor actor;
     private final Commander commander;
+    private final Gson gson;
 
     @Autowired
-    public CallbackApiHandler(VkMessageSenderService vkMessageSenderService, Commander commander) {
-        this.vkMessageSenderService = vkMessageSenderService;
+    public CallbackApiHandler(VKGroupActor actor, Commander commander) {
+        this.actor = actor;
         this.commander = commander;
+        this.gson = new Gson();
     }
 
     @Override
-    public void messageNew(Integer groupId, Message message) {
-        commander.execute(message);
-    }
-
-    @Override
-    public void confirmation(Integer groupId) {
-        Message message1 = new Message();
-        super.confirmation(groupId);
-        message1.setText("d70aff5a");
-        try {
-            vkMessageSenderService.send(message1);
-        } catch (ClientException e) {
-            e.printStackTrace();
-        } catch (ApiException e) {
-            e.printStackTrace();
+    public void messageNew(Integer groupId, String secret, Message message) {
+        if (secretIsValid(secret)){
+            commander.execute(message);
         }
-
     }
-    Gson gson = new Gson();
+
+    private boolean secretIsValid(String secret){
+        return secret.equals(actor.getSecret());
+    }
 
     @Override
     public boolean parse(String json) {
@@ -53,20 +40,23 @@ public class CallbackApiHandler extends CallbackApi {
         return parse(jsonObject);
     }
 
-    private static final String CALLBACK_EVENT_CONFIRMATION = "confirmation";
+
     @Override
     public boolean parse(JsonObject json) {
         String type = json.get("type").getAsString();
+        String CALLBACK_EVENT_CONFIRMATION = "confirmation";
+        String CALLBACK_EVENT_MESSAGE_NEW = "message_new";
         if (type.equalsIgnoreCase(CALLBACK_EVENT_CONFIRMATION)) {
             ConfirmationMessage message = gson.fromJson(json, ConfirmationMessage.class);
             confirmation(message.getGroupId(), message.getSecret());
             return true;
-        }else {
+        }else if (type.equalsIgnoreCase(CALLBACK_EVENT_MESSAGE_NEW)){
             int groupId = json.get("group_id").getAsInt();
             String s = json.get("secret").getAsString();
             Message message =  gson.fromJson(json.get("object").getAsJsonObject().get("message"),Message.class);
             messageNew(groupId, s, message);
             return true;
         }
+        return false;
     }
 }
