+++ src/main/resources/application.properties
@@ -10,7 +10,7 @@ spring.datasource.username=sa
 spring.datasource.password=
 
 vk.api.accesstoken=${ACCESS-TOKEN}
-vk.api.v=5.92
+vk.api.v=5.126
 vk.api.secret=${SECRET}
 vk.api.confirmation=${CONFIRMATION}
 vk.api.groupid=193662754
