--- src/main/resources/vk.properties
@@ -1,4 +1,0 @@
-vk.api.access-token=${ACCESS-TOKEN}
-vk.api.v=5.92
-vk.api.secret=${SECRET}
-vk.api.confirmation=${CONFIRMATION}
